converter/requirements.txt
pika
python-dotenv
websocket-client
fastapi
dependency_injector
uvicorn
redis